Split api token logic into MVC modules
