+++ src/controllers/mod.rs
@@ -19,3 +19,4 @@ pub mod crate_owner_invitation;
 pub mod keyword;
 pub mod krate;
 pub mod site_metadata;
+pub mod token;
